Update 13 + Beta 0.5 (dev)
1. Добавлены уровни сложности (Easy, Medium, Hard, Insane, Safe)
2. Удалена модель игрока
3. Добавлено попадание в голову
4. Добавлено отображение нанесённого урона

+++ Assets/Scripts/Game/Enemy/EnemyInfo.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class EnemyInfo : MonoBehaviour
 {
-    public int _maxHP;
-    [HideInInspector] public int _currentHP;
+    public float _maxHP;
+    [HideInInspector] public float _currentHP;
     public bool _agressive;
     public float _rangeAttack;
     public float _damage;
